requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3
pandas==2.2.0
tweepy==4.14.0
tf-keras==2.15.0
sentence-transformers==2.2.2
transformers==4.36.2